+++ src/service/atcoder.rs
@@ -20,7 +20,6 @@ use failure::ResultExt;
 use itertools::Itertools;
 use maplit::hashmap;
 use once_cell::sync::Lazy;
-use once_cell::sync_lazy;
 use regex::Regex;
 use reqwest::{header, StatusCode};
 use select::document::Document;
@@ -31,7 +30,6 @@ use tokio::runtime::{Runtime, TaskExecutor};
 
 use std::borrow::Cow;
 use std::collections::{BTreeMap, HashMap, HashSet};
-use std::ffi::OsStr;
 use std::io::Write;
 use std::ops::Deref;
 use std::path::Path;
@@ -657,13 +655,6 @@ impl<T: Term> Atcoder<T> {
                 let source_code = crate::fs::read_to_string(src_path)?;
                 let document = self.get(&url).recv_html()?;
                 let csrf_token = document.extract_csrf_token()?;
-                let lang_id = match lang_id.as_ref() {
-                    None => {
-                        let ext = src_path.extension().unwrap_or_default();
-                        Cow::from(document.extract_lang_id_by_extension(ext)?)
-                    }
-                    Some(lang_id) => Cow::from(lang_id.as_str()),
-                };
                 let url = contest.url_submit();
                 let payload = hashmap!(
                     "data.TaskScreenName" => task_screen_name.as_str(),
@@ -674,7 +665,7 @@ impl<T: Term> Atcoder<T> {
 
                 let error = |status: StatusCode, location: Option<String>| -> _ {
                     ServiceError::from(ServiceErrorKind::SubmissionRejected(
-                        lang_id.as_ref().to_owned(),
+                        lang_id.clone(),
                         source_code.len(),
                         status,
                         location,
@@ -878,7 +869,6 @@ trait Extract {
     fn extract_submissions(&self) -> ScrapeResult<(vec::IntoIter<Submission>, u32)>;
     fn extract_submitted_code(&self) -> ScrapeResult<String>;
     fn extract_lang_id_by_name(&self, lang_name: &str) -> ScrapeResult<String>;
-    fn extract_lang_id_by_extension(&self, ext: &OsStr) -> ServiceResult<String>;
 }
 
 impl Extract for Document {
@@ -1233,132 +1223,6 @@ impl Extract for Document {
             }
         }
         Err(ScrapeError::new())
-    }
-
-    fn extract_lang_id_by_extension(&self, ext: &OsStr) -> ServiceResult<String> {
-        enum Kind {
-            Mime(&'static str),
-            Name(&'static str),
-            Ambiguous(&'static [&'static str]),
-        }
-
-        static KINDS: Lazy<HashMap<&OsStr, Kind>> = sync_lazy!(hashmap!(
-            OsStr::new("bash")   => Kind::Name("Bash"),
-            OsStr::new("sh")     => Kind::Name("Bash"),
-            OsStr::new("c")      => Kind::Mime("text/x-csrc"),
-            OsStr::new("cpp")    => Kind::Mime("text/x-c++src"),
-            OsStr::new("cxx")    => Kind::Mime("text/x-c++src"),
-            OsStr::new("cc")     => Kind::Mime("text/x-c++src"),
-            OsStr::new("C")      => Kind::Mime("text/x-c++src"),
-            OsStr::new("cs")     => Kind::Mime("text/x-csharp"),
-            OsStr::new("clj")    => Kind::Mime("text/x-closure"),
-            OsStr::new("lisp")   => Kind::Mime("text/x-common-lisp"),
-            OsStr::new("cl")     => Kind::Mime("text/x-common-lisp"),
-            OsStr::new("d")      => Kind::Mime("text/x-d"),
-            OsStr::new("f08")    => Kind::Mime("text/x-fortran"),
-            OsStr::new("F08")    => Kind::Mime("text/x-fortran"),
-            OsStr::new("f03")    => Kind::Mime("text/x-fortran"),
-            OsStr::new("F03")    => Kind::Mime("text/x-fortran"),
-            OsStr::new("f95")    => Kind::Mime("text/x-fortran"),
-            OsStr::new("F95")    => Kind::Mime("text/x-fortran"),
-            OsStr::new("f90")    => Kind::Mime("text/x-fortran"),
-            OsStr::new("F90")    => Kind::Mime("text/x-fortran"),
-            OsStr::new("f")      => Kind::Mime("text/x-fortran"),
-            OsStr::new("for")    => Kind::Mime("text/x-fortran"),
-            OsStr::new("go")     => Kind::Mime("text/x-go"),
-            OsStr::new("hs")     => Kind::Mime("text/x-haskell"),
-            OsStr::new("java")   => Kind::Mime("text/x-java"),
-            OsStr::new("js")     => Kind::Mime("text/javascript"),
-            OsStr::new("ml")     => Kind::Mime("text/x-ocaml"),
-            OsStr::new("pas")    => Kind::Mime("text/x-pascal"),
-            OsStr::new("pl")     => Kind::Mime("text/x-perl"),
-            OsStr::new("php")    => Kind::Mime("text/x-php"),
-            OsStr::new("py")     => Kind::Mime("text/x-python"),
-            OsStr::new("py2")    => Kind::Mime("text/x-python"),
-            OsStr::new("py3")    => Kind::Mime("text/x-python"),
-            OsStr::new("rb")     => Kind::Mime("text/x-ruby"),
-            OsStr::new("scala")  => Kind::Mime("text/x-scala"),
-            OsStr::new("scm")    => Kind::Mime("text/x-scheme"),
-            OsStr::new("txt")    => Kind::Mime("text/plain"),
-            OsStr::new("vb")     => Kind::Mime("text/x-vb"),
-            OsStr::new("m")      => Kind::Ambiguous(&["Objective-C", "Octave"]),
-            OsStr::new("swift")  => Kind::Mime("text/x-swift"),
-            OsStr::new("rs")     => Kind::Mime("text/x-rust"),
-            OsStr::new("sed")    => Kind::Name("Sed"),
-            OsStr::new("awk")    => Kind::Name("Awk"),
-            OsStr::new("bf")     => Kind::Mime("text/x-brainfuck"),
-            OsStr::new("sml")    => Kind::Mime("text/x-sml"),
-            OsStr::new("cr")     => Kind::Mime("text/x-crystal"),
-            OsStr::new("fs")     => Kind::Mime("text/x-fsharp"),
-            OsStr::new("unl")    => Kind::Mime("text/x-unlambda"),
-            OsStr::new("lua")    => Kind::Mime("text/x-lua"),
-            OsStr::new("moon")   => Kind::Mime("text/x-moonscript"),
-            OsStr::new("ceylon") => Kind::Mime("text/x-ceylon"),
-            OsStr::new("jl")     => Kind::Mime("text/x-julia"),
-            OsStr::new("nim")    => Kind::Mime("text/x-nim"),
-            OsStr::new("ts")     => Kind::Mime("text/typescript"),
-            OsStr::new("p6")     => Kind::Name("Perl6"),
-            OsStr::new("kt")     => Kind::Mime("text/x-kotlin"),
-            OsStr::new("cob")    => Kind::Mime("text/x-cobol"),
-        ));
-        static NAME: Lazy<Regex> = lazy_regex!(r#"\A(.*?)\s*\(.*?\)\z"#);
-
-        macro_rules! with_msg {
-            ($msg:expr) => {
-                ServiceError::from(failure::err_msg($msg).context(
-                    ServiceErrorKind::RecognizeByExtension(ext.to_string_lossy().into_owned()),
-                ))
-            };
-        }
-
-        let kind = KINDS
-            .get(ext)
-            .ok_or_else(|| with_msg!("Unknown extension"))?;
-        if let Kind::Ambiguous(candidates) = kind {
-            return Err(with_msg!(format!(
-                "Ambiguous (candidates: {{ {} }})",
-                candidates
-                    .iter()
-                    .format_with(", ", |s, f| f(&format_args!("????: {:?} (???)", s)))
-            )));
-        }
-        let mut matched = vec![];
-        for option in self.find(selector!("#select-lang > select > option")) {
-            let lang_id = option.attr("value").ok_or_else(ScrapeError::new)?;
-            let mime = option.attr("data-mime").ok_or_else(ScrapeError::new)?;
-            let name = option
-                .find(Text)
-                .next()
-                .ok_or_else(ScrapeError::new)?
-                .text();
-            match kind {
-                Kind::Ambiguous(_) => unreachable!(),
-                Kind::Mime(s) => {
-                    if s == &mime {
-                        matched.push((lang_id, name));
-                    }
-                }
-                Kind::Name(s) => {
-                    let p = {
-                        let caps = NAME.captures(&name).ok_or_else(ScrapeError::new)?;
-                        s == &&caps[1]
-                    };
-                    if p {
-                        matched.push((lang_id, name));
-                    }
-                }
-            }
-        }
-        if matched.len() == 1 {
-            Ok(matched[0].0.to_owned())
-        } else {
-            Err(with_msg!(format!(
-                "Candidates:\n{}",
-                matched
-                    .iter()
-                    .format_with("", |(n, s), f| f(&format_args!("  {}: {:?}\n", n, s)))
-            )))
-        }
     }
 }