+++ src/service/yukicoder.rs
@@ -1,4 +1,4 @@
-use crate::errors::{ScrapeError, ScrapeResult, ServiceError, ServiceErrorKind, ServiceResult};
+use crate::errors::{ScrapeError, ScrapeResult, ServiceErrorKind, ServiceResult};
 use crate::service::download::DownloadProgress;
 use crate::service::session::HttpSession;
 use crate::service::{
@@ -13,7 +13,6 @@ use crate::util::str::CaseConversion;
 use cookie::Cookie;
 use failure::ResultExt;
 use itertools::Itertools;
-use maplit::hashmap;
 use once_cell::sync::Lazy;
 use once_cell::sync_lazy;
 use regex::Regex;
@@ -24,8 +23,6 @@ use serde_derive::Deserialize;
 use tokio::runtime::{Runtime, TaskExecutor};
 
 use std::borrow::Cow;
-use std::collections::HashMap;
-use std::ffi::OsStr;
 use std::io::Write;
 use std::str::FromStr;
 use std::time::Duration;
@@ -307,58 +304,6 @@ impl<T: Term> Yukicoder<T> {
     }
 
     fn submit(&mut self, props: &SubmitProps<YukicoderContest>) -> ServiceResult<()> {
-        static LANG_IDS: Lazy<HashMap<&OsStr, &[&str]>> = sync_lazy!(hashmap!(
-            OsStr::new("cpp")   => ["cpp", "cpp14", "cpp17", "cpp-clang"].as_ref(),
-            OsStr::new("cxx")   => &["cpp", "cpp14", "cpp17", "cpp-clang"],
-            OsStr::new("cc")    => &["cpp", "cpp14", "cpp17", "cpp-clang"],
-            OsStr::new("C")     => &["cpp", "cpp14", "cpp17", "cpp-clang"],
-            OsStr::new("c")     => &["c11", "c"],
-            OsStr::new("java")  => &["java8"],
-            OsStr::new("cs")    => &["csharp", "csharp_mono"],
-            OsStr::new("pl")    => &["perl", "perl6"],
-            OsStr::new("p6")    => &["perl6"],
-            OsStr::new("php")   => &["php", "php7"],
-            OsStr::new("py")    => &["python", "python3", "pypy2", "pypy3"],
-            OsStr::new("py2")   => &["python", "pypy2"],
-            OsStr::new("py3")   => &["python3", "pypy3"],
-            OsStr::new("rb")    => &["ruby"],
-            OsStr::new("d")     => &["d"],
-            OsStr::new("go")    => &["go"],
-            OsStr::new("hs")    => &["haskell"],
-            OsStr::new("scala") => &["scala"],
-            OsStr::new("nim")   => &["nim"],
-            OsStr::new("rs")    => &["rust"],
-            OsStr::new("kt")    => &["kotlin"],
-            OsStr::new("scm")   => &["scheme"],
-            OsStr::new("cr")    => &["crystal"],
-            OsStr::new("swift") => &["swift"],
-            OsStr::new("ml")    => &["ocaml"],
-            OsStr::new("clj")   => &["clojure"],
-            OsStr::new("fs")    => &["fsharp"],
-            OsStr::new("exs")   => &["elixer"],
-            OsStr::new("ex")    => &["elixer"],
-            OsStr::new("lua")   => &["lua"],
-            OsStr::new("f")     => &["fortran"],
-            OsStr::new("for")   => &["fortran"],
-            OsStr::new("f90")   => &["fortran"],
-            OsStr::new("F90")   => &["fortran"],
-            OsStr::new("f95")   => &["fortran"],
-            OsStr::new("F95")   => &["fortran"],
-            OsStr::new("f03")   => &["fortran"],
-            OsStr::new("F03")   => &["fortran"],
-            OsStr::new("f08")   => &["fortran"],
-            OsStr::new("F08")   => &["fortran"],
-            OsStr::new("js")    => &["node"],
-            OsStr::new("vim")   => &["vim"],
-            OsStr::new("sh")    => &["sh"],
-            OsStr::new("bash")  => &["sh"],
-            OsStr::new("txt")   => &["text"],
-            OsStr::new("asm")   => &["nasm"],
-            OsStr::new("clay")  => &["clay"], // ?
-            OsStr::new("bf")    => &["bf"],
-            OsStr::new("ws")    => &["Whitespace"],
-        ));
-
         let SubmitProps {
             contest,
             problem,
@@ -368,29 +313,6 @@ impl<T: Term> Yukicoder<T> {
             skip_checking_if_accepted,
         } = props;
 
-        let lang_id = match lang_id {
-            None => {
-                let ext = src_path.extension().unwrap_or_default();
-                let error = |e: failure::Error| -> ServiceError {
-                    let ext = ext.to_string_lossy().into_owned();
-                    e.context(ServiceErrorKind::RecognizeByExtension(ext))
-                        .into()
-                };
-                match LANG_IDS.get(ext) {
-                    Some(&[id]) => Cow::from(*id),
-                    Some(ids) => {
-                        let msg = format!(
-                            "Candidates: [{}]",
-                            ids.iter()
-                                .format_with(", ", |s, f| f(&format_args!("{:?}", s))),
-                        );
-                        return Err(error(failure::err_msg(msg)));
-                    }
-                    None => return Err(error(failure::err_msg("Unknown extension"))),
-                }
-            }
-            Some(lang_id) => Cow::from(lang_id.as_str()),
-        };
         let code = crate::fs::read_to_string(src_path)?;
 
         self.login(true)?;
@@ -421,7 +343,7 @@ impl<T: Term> Yukicoder<T> {
         let token = document.extract_csrf_token_from_submit_page()?;
         let form = reqwest::r#async::multipart::Form::new()
             .text("csrf_token", token)
-            .text("lang", lang_id.clone().into_owned())
+            .text("lang", lang_id.clone())
             .text("source", code.clone());
         let url = document.extract_url_from_submit_page()?;
         let res = self.post(&url).send_multipart(form)?;
@@ -444,7 +366,7 @@ impl<T: Term> Yukicoder<T> {
             }
         }
         Err(ServiceErrorKind::SubmissionRejected(
-            lang_id.as_ref().to_owned(),
+            lang_id.clone(),
             code.len(),
             res.status(),
             location.map(ToOwned::to_owned),